RENAME db_module -> persistence
